+++ platforms/react-native/src/theme/index.ts
@@ -1,6 +1,11 @@
 import { makeTheme } from '@shallot-ui/core-theme'
-import nativeVariants from './variants'
+import nativeThemeVariants from './variants'
+import nativeThemeGlobals from './globals'
 
-export { nativeVariants }
-export const DEFAULT_NATIVE_THEME = makeTheme({}, nativeVariants)
+export { nativeThemeVariants, nativeThemeGlobals }
+export const DEFAULT_NATIVE_THEME = makeTheme(
+  {},
+  nativeThemeVariants,
+  nativeThemeGlobals,
+)
 export type NativeTheme = typeof DEFAULT_NATIVE_THEME