+++ platforms/react-native/src/theme/variants/Text.ts
@@ -1,27 +1,35 @@
-import type { ShallotProp } from '@shallot-ui/core-theme'
-import {
-  getFontFamily,
-  getFontSize,
-  getLineHeight,
-} from '@shallot-ui/core-mixins'
+import { getFontFamily, getFontSize, getGlobal } from '@shallot-ui/core-mixins'
+import { ShallotProp } from '@shallot-ui/core-theme'
 
 const Text: Record<string, ShallotProp> = {
+  Default: {
+    fontFamily: getFontFamily('Body'),
+    fontSize: getFontSize('md'),
+    color: getGlobal('foregroundColor'),
+    fontWeight: 300,
+  },
+  P: {
+    fontFamily: getFontFamily('Body'),
+    fontSize: getFontSize('md'),
+    color: getGlobal('foregroundColor'),
+    fontWeight: 300,
+  },
   H1: {
     fontFamily: getFontFamily('Body'),
     fontSize: getFontSize('xl'),
-    lineHeight: getLineHeight('xl'),
+    color: getGlobal('foregroundFadeColor'),
     fontWeight: 600,
   },
   H2: {
     fontFamily: getFontFamily('Body'),
     fontSize: getFontSize('lg'),
-    lineHeight: getLineHeight('lg'),
+    color: getGlobal('foregroundFadeColor'),
     fontWeight: 600,
   },
   H3: {
     fontFamily: getFontFamily('Body'),
     fontSize: getFontSize('md'),
-    lineHeight: getLineHeight('md'),
+    color: getGlobal('foregroundColor'),
     fontWeight: 600,
   },
 }