+++ platforms/react/src/index.ts
@@ -1,5 +1,6 @@
 export * from './hooks'
 export * from './components'
+export * from './context'
 export * from './styles'
 export * from './utils'
 export * from './theme'