+++ platforms/react/styled.d.ts
@@ -1,7 +1,7 @@
-import { DEFAULT_WEB_THEME } from './src'
+import { DEFAULT_REACT_THEME } from './src'
 import 'styled-components'
 
-type Theme = typeof DEFAULT_WEB_THEME
+type Theme = typeof DEFAULT_REACT_THEME
 
 declare module 'styled-components' {
   export interface DefaultTheme extends Theme {}