+++ components/Tag/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @shallot-ui/component-tag
 
+## 3.0.0-beta.12
+
+### Patch Changes
+
+- Generalizing providers
+- Updated dependencies
+  - @shallot-ui/core-mixins@2.0.0-beta.10
+  - @shallot-ui/core-theme@2.0.0-beta.10
+  - @shallot-ui/core-utils@2.0.0-beta.11
+
 ## 3.0.0-beta.11
 
 ### Patch Changes